Consolidated the exception classes and removed 'cause' from the error message
