+++ mora/auth.py
@@ -20,7 +20,7 @@ import os
 import flask
 import requests.auth
 
-from .errorcodes import ErrorCodes
+from mora.exceptions import ErrorCodes
 from . import exceptions
 from . import tokens
 
@@ -88,7 +88,7 @@ def login(username):
             flask.request.full_path,
         )
 
-        raise exceptions.UnauthorizedError(ErrorCodes.E_CONNECTION_FAILED)
+        raise exceptions.BaseError(ErrorCodes.E_CONNECTION_FAILED)
 
     resp = flask.jsonify({
         "user": username,
