+++ mora/lora.py
@@ -10,13 +10,13 @@ from __future__ import generator_stop
 
 import collections
 import datetime
-import itertools
 import functools
+import itertools
 import uuid
 
 import requests
 
-from .errorcodes import ErrorCodes
+from mora.exceptions import ErrorCodes
 from . import auth
 from . import exceptions
 from . import settings
@@ -153,12 +153,15 @@ def _check_response(r):
         try:
             d = r.json()
         except ValueError:
-            raise exceptions.ValidationError(r.text)
+            raise exceptions.BaseError(ErrorCodes.E_INVALID_INPUT,
+                                       message=r.text)
 
         if r.status_code == 400 and d:
-            raise exceptions.ValidationError(message=r.json()['message'])
+            raise exceptions.BaseError(ErrorCodes.E_INVALID_INPUT,
+                                       message=r.json()['message'])
         elif r.status_code in (401, 403) and d:
-            raise exceptions.UnauthorizedError(message=r.json()['message'])
+            raise exceptions.BaseError(ErrorCodes.E_UNAUTHORIZED,
+                                       message=r.json()['message'])
         else:
             raise
 
@@ -267,7 +270,7 @@ class Connector:
             self.__daterange = (self.today, self.tomorrow)
 
         else:
-            raise exceptions.ValidationError(
+            raise exceptions.BaseError(
                 ErrorCodes.V_INVALID_VALIDITY,
                 validity=self.__validity
             )
