+++ mora/service/address.py
@@ -154,7 +154,7 @@ import re
 import flask
 import requests
 
-from ..errorcodes import ErrorCodes
+from mora.exceptions import ErrorCodes
 from .. import exceptions
 from .. import lora
 from .. import util
@@ -231,7 +231,7 @@ def get_relation_for(addrobj, fallback=None):
         r['objekttype'] = common.get_uuid(typeobj)
 
     else:
-        raise exceptions.ValidationError(
+        raise exceptions.BaseError(
             'unknown address scope {!r}!'.format(scope),
         )
 
@@ -289,7 +289,7 @@ def get_one_address(c, addrrel, class_cache=None):
         urn = addrrel['urn']
 
         if not urn.startswith(prefix):
-            raise exceptions.ValidationError('invalid urn {!r}'.format(
+            raise exceptions.BaseError('invalid urn {!r}'.format(
                 addrrel['urn'],
             ))
 
@@ -313,7 +313,7 @@ def get_one_address(c, addrrel, class_cache=None):
         }
 
     else:
-        raise exceptions.ValidationError(
+        raise exceptions.BaseError(
             'invalid address scope {!r}'.format(addrformat),
         )
 
@@ -415,7 +415,7 @@ class Addresses(common.AbstractRelationDetail):
         new_entry = common.checked_get(req, 'data', {}, required=True)
 
         if not old_entry:
-            raise exceptions.ValidationError(
+            raise exceptions.BaseError(
                 ErrorCodes.V_ORIGINAL_REQUIRED
             )
 
@@ -425,7 +425,7 @@ class Addresses(common.AbstractRelationDetail):
         try:
             addresses = original['relationer']['adresser']
         except KeyError:
-            raise exceptions.ValidationError('no addresses to edit!')
+            raise exceptions.BaseError('no addresses to edit!')
 
         addresses = common.replace_relation_value(addresses, old_rel, new_rel)
 
@@ -481,7 +481,7 @@ def address_autocomplete(orgid):
         org = lora.Connector().organisation.get(orgid)
 
         if not org:
-            raise exceptions.NotFoundError(ErrorCodes.E_NO_LOCAL_MUNICIPALITY)
+            raise exceptions.BaseError(ErrorCodes.E_NO_LOCAL_MUNICIPALITY)
 
         for myndighed in org.get('relationer', {}).get('myndighed', []):
             m = MUNICIPALITY_CODE_PATTERN.fullmatch(myndighed.get('urn'))
@@ -490,7 +490,7 @@ def address_autocomplete(orgid):
                 code = int(m.group(1))
                 break
         else:
-            raise exceptions.NotFoundError(ErrorCodes.E_NO_LOCAL_MUNICIPALITY)
+            raise exceptions.BaseError(ErrorCodes.E_NO_LOCAL_MUNICIPALITY)
     else:
         code = None
