+++ mora/service/cpr.py
@@ -15,7 +15,7 @@ based on their CPR number.
 
 import flask
 
-from ..errorcodes import ErrorCodes
+from mora.exceptions import ErrorCodes
 from mora import exceptions
 from mora import util
 from mora.integrations.serviceplatformen import get_citizen
@@ -52,12 +52,12 @@ def search_cpr():
     try:
         sp_data = get_citizen(cpr)
     except KeyError:
-        raise exceptions.NotFoundError(
+        raise exceptions.BaseError(
             ErrorCodes.V_NO_PERSON_FOR_CPR,
             cpr=cpr,
         )
     except ValueError:
-        raise exceptions.ValidationError(
+        raise exceptions.BaseError(
             ErrorCodes.V_CPR_NOT_VALID,
             cpr=cpr,
         )
