+++ mora/service/facet.py
@@ -131,7 +131,7 @@ def get_one_class(c, classid, clazz=None):
         clazz = c.klasse.get(classid)
 
         if not clazz:
-            raise exceptions.NotFoundError(
+            raise exceptions.BaseError(
                 'no such class {!r}'.format(classid),
             )
