+++ mora/service/itsystem.py
@@ -21,7 +21,7 @@ import uuid
 import flask
 import werkzeug
 
-from ..errorcodes import ErrorCodes
+from mora.exceptions import ErrorCodes
 from .. import exceptions
 from .. import util
 
@@ -241,7 +241,7 @@ class ITSystems(common.AbstractRelationDetail):
         )
 
         if not original:
-            raise exceptions.NotFoundError(ErrorCodes.E_USER_NOT_FOUND)
+            raise exceptions.BaseError(ErrorCodes.E_USER_NOT_FOUND)
 
         rels = original['relationer'].get('tilknyttedeitsystemer', [])
 
@@ -270,7 +270,7 @@ class ITSystems(common.AbstractRelationDetail):
         old_rel = original['relationer'].get('tilknyttedeitsystemer', [])
 
         if not old_entry:
-            raise exceptions.ValidationError(ErrorCodes.V_ORIGINAL_REQUIRED)
+            raise exceptions.BaseError(ErrorCodes.V_ORIGINAL_REQUIRED)
 
         # We are performing an update of a pre-existing effect
         old_rel = self.get_relation_for(
