+++ mora/service/orgunit.py
@@ -22,7 +22,7 @@ import uuid
 import flask
 import werkzeug
 
-from ..errorcodes import ErrorCodes
+from mora.exceptions import ErrorCodes
 from . import address
 from . import common
 from . import facet
@@ -211,7 +211,7 @@ def get_one_orgunit(c, unitid, unit=None,
         )
 
     else:
-        raise exceptions.ValidationError(
+        raise exceptions.BaseError(
             'invalid details {!r}'.format(details),
         )
 
@@ -715,7 +715,7 @@ def terminate_org_unit(unitid):
     c = lora.Connector(effective_date=date)
 
     if not c.organisationenhed.get(unitid):
-        raise exceptions.NotFoundError(ErrorCodes.E_ORG_UNIT_NOT_FOUND)
+        raise exceptions.BaseError(ErrorCodes.E_ORG_UNIT_NOT_FOUND)
 
     children = c.organisationenhed.paged_get(
         get_one_orgunit,
@@ -730,7 +730,7 @@ def terminate_org_unit(unitid):
     )
 
     if children['total'] or roles:
-        raise exceptions.ValidationError(
+        raise exceptions.BaseError(
             ErrorCodes.V_TERMINATE_UNIT_WITH_CHILDREN_OR_ROLES,
             child_units=children['items'],
             child_count=children['total'],
