+++ mora/tokens.py
@@ -14,6 +14,7 @@ import lxml.etree
 import flask
 import requests
 
+from mora.exceptions import ErrorCodes
 from . import exceptions
 from . import settings
 from . import util
@@ -95,7 +96,7 @@ def get_token(username, passwd, raw=False, verbose=False, insecure=None):
                             None, XML_NAMESPACES)
 
     if not resp.ok or errormsg:
-        raise exceptions.UnauthorizedError(message=errormsg)
+        raise exceptions.BaseError(ErrorCodes.E_UNAUTHORIZED, message=errormsg)
 
     tokens = doc.findall('.//saml:Assertion', XML_NAMESPACES)
