+++ mora/util.py
@@ -22,7 +22,7 @@ import flask
 import dateutil.parser
 import dateutil.tz
 
-from .errorcodes import ErrorCodes
+from mora.exceptions import ErrorCodes
 from . import exceptions
 
 
@@ -93,7 +93,7 @@ def parsedatetime(s: str) -> datetime.datetime:
     try:
         dt = dateutil.parser.parse(s, dayfirst=True, tzinfos=tzinfos)
     except ValueError:
-        raise exceptions.ValidationError('cannot parse {!r}'.format(s))
+        raise exceptions.BaseError('cannot parse {!r}'.format(s))
 
     return dt
 
@@ -258,7 +258,7 @@ def update_config(mapping, config_path, allow_environment=True):
 
 def splitlist(xs, size):
     if size <= 0:
-        raise exceptions.ValidationError(ErrorCodes.E_SIZE_MUST_BE_POSITIVE)
+        raise exceptions.BaseError(ErrorCodes.E_SIZE_MUST_BE_POSITIVE)
 
     i = 0
     nxs = len(xs)
