+++ mora/validator.py
@@ -9,7 +9,7 @@ import datetime
 from mora import exceptions
 from mora import lora
 from mora import util
-from .errorcodes import ErrorCodes
+from mora.exceptions import ErrorCodes
 
 
 def _is_date_range_valid(parent: str, startdate: datetime.datetime,
@@ -70,7 +70,7 @@ def is_date_range_in_org_unit_range(org_unit_uuid, valid_from, valid_to):
 
     if not _is_date_range_valid(org_unit_uuid, valid_from, valid_to, scope,
                                 gyldighed_key):
-        raise exceptions.ValidationError(
+        raise exceptions.BaseError(
             ErrorCodes.V_DATE_OUTSIDE_ORG_UNIT_RANGE,
             org_unit_uuid=org_unit_uuid,
             valid_from=valid_from,
@@ -87,7 +87,7 @@ def is_date_range_in_employee_range(employee_uuid, valid_from, valid_to):
 
     if not _is_date_range_valid(employee_uuid, valid_from, valid_to, scope,
                                 gyldighed_key):
-        raise exceptions.ValidationError(
+        raise exceptions.BaseError(
             ErrorCodes.V_DATE_OUTSIDE_EMPL_RANGE,
             employee_uuid=employee_uuid,
             valid_from=valid_from,
@@ -115,7 +115,7 @@ def is_candidate_parent_valid(old_unitid: str, new_unitid: str,
     )['relationer']
     if org_unit_relations['overordnet'][0]['uuid'] == \
             org_unit_relations['tilhoerer'][0]['uuid']:
-        raise exceptions.ValidationError(
+        raise exceptions.BaseError(
             ErrorCodes.V_CANNOT_MOVE_ROOT_ORG_UNIT)
 
     # Use for checking that the candidate parent is not the units own subtree
@@ -141,7 +141,7 @@ def is_candidate_parent_valid(old_unitid: str, new_unitid: str,
         return is_node_valid(parent)
 
     if not is_node_valid(new_unitid):
-        raise exceptions.ValidationError(ErrorCodes.V_ORG_UNIT_MOVE_TO_CHILD)
+        raise exceptions.BaseError(ErrorCodes.V_ORG_UNIT_MOVE_TO_CHILD)
 
 
 def _get_org_unit_endpoint_date(org_unit: dict,
@@ -162,7 +162,7 @@ def _get_org_unit_endpoint_date(org_unit: dict,
             else:
                 return util.parsedatetime(virkning['from'])
 
-    raise exceptions.ValidationError('the unit did not have an end date!')
+    raise exceptions.BaseError('the unit did not have an end date!')
 
 
 def is_inactivation_date_valid(unitid: str, end_date: str) -> bool:
